+++ src/containers/ModuleList.js
@@ -1,7 +1,8 @@
 import React, {Component} from 'react';
+import {BrowserRouter as Router, Route} from 'react-router-dom'
 import ModuleListItem from '../components/ModuleListItem';
 import ModuleServiceClient from '../services/ModuleServiceClient';
-
+import ModuleEditor from "./ModuleEditor";
 
 export default class ModuleList extends Component {
     constructor(props) {
@@ -11,7 +12,7 @@ export default class ModuleList extends Component {
             module: { title: '' },
             modules: [
 
-            ]
+            ],
         };
         this.createModule = this.createModule.bind(this);
         this.titleChanged = this.titleChanged.bind(this);
@@ -35,10 +36,10 @@ export default class ModuleList extends Component {
     componentDidMount() {
         this.setCourseId(this.props.courseId);
     }
-    componentWillReceiveProps(newProps){
-        this.setCourseId(newProps.courseId);
-        this.findAllModulesForCourse(newProps.courseId)
-    }
+     componentWillReceiveProps(newProps){
+          this.setCourseId(newProps.courseId);
+          this.findAllModulesForCourse(newProps.courseId);
+      }
 
     createModule() {
         this.moduleService
@@ -51,53 +52,62 @@ export default class ModuleList extends Component {
     renderListOfModules() {
         let modules = null;
         if(this.state) {
-            modules = this.state.modules.map(
-                function (module) {
-                    console.log(module);
+            modules = this.state.modules.map((module) =>{
                     return <ModuleListItem key={module.id}
-                                           module={module}/>
+                                           module={module}
+                                           courseId = {this.props.courseId}
+                                           deleteModule={this.deleteModule}/>
                 }
-            )
-            modules = this.state.modules.map((module) => {
-                    return <ModuleListItem module={module}
-                                           key={module.id}
-                                           deleteModule={this.deleteModule}
-                                           loadLessons ={this.loadLessons}/>
-            });
+            );
         }
-
         return (
             modules
         )
     }
     loadLessons(moduleId){
         this.props.loadLessons(moduleId);
+        this.setState({active: {moduleId: moduleId ,value:true}});
     }
     deleteModule(moduleId) {
-        this.moduleService
-            .deleteModule(moduleId)
-            .then(() => { this.findAllModulesForCourse(this.props.courseId); });
+        let value= window.confirm(`Deleting the module`);
+        if(value) {
+            this.moduleService
+                .deleteModule(moduleId)
+                .then(() => {
+                    this.findAllModulesForCourse(this.props.courseId);
+                });
+        }
     }
     render() {
-        return <div>
-            <ul className="list-group ">
-                <li className="list-group-item" style={{backgroundColor: "#F6BB42" }}>
-                    <div className="form-group-inline">
-                        <div className="input-group">
-                            <input type="text" placeholder="Title" className="form-control"
-                                       onChange={this.titleChanged}
-                                       value={this.state.module.title}/>
-                            <div className="input-group-btn">
-                                <button type="button" className="btn btn-dark">
-                                    <i className="fa fa-plus inline"
-                                           onClick={this.createModule}/>
-                                </button>
-                            </div>
+        return (<div>
+                <Router>
+                    <div className="row">
+                        <div className="col-4">
+                            <ul className="list-group ">
+                                <li className="list-group-item" style={{backgroundColor: "#F6BB42" }}>
+                                    <div className="form-group-inline">
+                                        <div className="input-group">
+                                            <input type="text" placeholder="Module Title" className="form-control"
+                                                   onChange={this.titleChanged}
+                                                   value={this.state.module.title}/>
+                                            <div className="input-group-btn">
+                                                <button type="button" className="btn btn-dark">
+                                                    <i className="fa fa-plus inline"
+                                                        onClick={this.createModule}/>
+                                                 </button>
+                                            </div>
+                                         </div>
+                                    </div>
+                                </li>
+                                {this.renderListOfModules()}
+                            </ul>
+                        </div>
+                        <div className="col-8">
+                            <Route path="/course/:courseId/module/:moduleId"
+                                    component = {ModuleEditor}/>
                         </div>
                     </div>
-                </li>
-                {this.renderListOfModules()}
-            </ul>
-        </div>;
+                </Router>
+        </div>);
     }
 }
